+++ ts-client/src/amm/types/index.ts
@@ -3,7 +3,7 @@ import { TokenInfo } from '@solana/spl-token-registry';
 import { IdlAccounts, IdlTypes, Program } from '@coral-xyz/anchor';
 import BN from 'bn.js';
 import { Amm as AmmIdl } from '../idl';
-import { VaultState, VaultIdl } from '@mercurial-finance/vault-sdk';
+import { VaultState, VaultIdl } from '@meteora-ag/vault-sdk';
 import Decimal from 'decimal.js';
 import { publicKey, struct, u64, u8, option, i64 } from '@coral-xyz/borsh';